requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
argon2-cffi=20.1.0=pypi_0
astroid=2.4.2=py37_0
async-generator=1.10=pypi_0
attrs=20.2.0=pypi_0
autopep8=1.5.4=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
bleach=3.2.1=pypi_0
ca-certificates=2020.10.14=0
certifi=2020.6.20=py37_0
cffi=1.14.3=pypi_0
chardet=3.0.4=pypi_0
click=7.1.2=pypi_0
configparser=5.0.1=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.3=pypi_0
filelock=3.0.12=pypi_0
future=0.18.2=pypi_0
gensim=3.8.3=pypi_0
gitdb=4.0.5=pypi_0
gitpython=3.1.11=pypi_0
idna=2.10=pypi_0
importlib-metadata=2.0.0=pypi_0
intel-openmp=2020.2=254
ipykernel=5.3.4=pypi_0
ipython=7.18.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
isort=5.6.4=py_0
jedi=0.17.2=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.17.0=py_0
jsonpickle=1.4.1=pypi_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.7=pypi_0
jupyter-console=6.2.0=pypi_0
jupyter-core=4.6.3=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
lazy-object-proxy=1.4.3=py37h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
markupsafe=1.1.1=pypi_0
mccabe=0.6.1=py37_1
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
nbclient=0.5.1=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.0.8=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.1=pypi_0
ninja=1.10.1=py37hfd86e86_0
nltk=3.5=pypi_0
notebook=6.1.4=pypi_0
numpy=1.19.2=pypi_0
numpy-base=1.19.1=py37hfa32c7d_0
oauthlib=3.1.0=pypi_0
openssl=1.1.1h=h7b6447c_0
packaging=20.4=pypi_0
pandas=1.1.3=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.7.1=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pip=20.2.4=py37_0
pip-chill=1.0.0=pypi_0
prometheus-client=0.8.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.8=pypi_0
protobuf=3.13.0=pypi_0
psutil=5.7.3=pypi_0
ptyprocess=0.6.0=pypi_0
pycodestyle=2.6.0=pypi_0
pycparser=2.20=pypi_0
pygments=2.7.1=pypi_0
pylint=2.6.0=py37_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
pysocks=1.7.1=pypi_0
python=3.7.9=h7579374_0
python-dateutil=2.8.1=pypi_0
pytorch=1.6.0=py3.7_cuda10.2.89_cudnn7.6.5_0
pytorch-nlp=0.5.0=pypi_0
pytz=2020.1=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.2=pypi_0
qtconsole=4.7.7=pypi_0
qtpy=1.9.0=pypi_0
readline=8.0=h7b6447c_0
regex=2020.10.23=pypi_0
requests=2.24.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
sacremoses=0.0.43=pypi_0
scikit-learn=0.23.2=py37h0573a6f_0
scipy=1.5.3=pypi_0
send2trash=1.5.0=pypi_0
sentencepiece=0.1.91=pypi_0
sentry-sdk=0.19.1=pypi_0
setuptools=50.3.0=py37hb0f4dca_1
shortuuid=1.0.1=pypi_0
six=1.15.0=py_0
smart-open=3.0.0=pypi_0
smmap=3.0.4=pypi_0
sqlite=3.33.0=h62c20be_0
subprocess32=3.5.4=pypi_0
terminado=0.9.1=pypi_0
testpath=0.4.4=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tokenizers=0.9.3=pypi_0
toml=0.10.1=py_0
torchtext=0.7.0=pypi_0
tornado=6.0.4=pypi_0
tqdm=4.50.2=pypi_0
traitlets=5.0.5=pypi_0
transformers=3.5.0=pypi_0
tweepy=3.9.0=pypi_0
typed-ast=1.4.1=py37h7b6447c_0
urllib3=1.25.11=pypi_0
wandb=0.10.8=pypi_0
watchdog=0.10.3=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.11.2=py37h7b6447c_0
xz=5.2.5=h7b6447c_0
zipp=3.3.1=pypi_0
zlib=1.2.11=h7b6447c_3
